+++ src/Sleet/CreateConfigAppCommand.cs
@@ -27,13 +27,15 @@ namespace Sleet
                 CommandOptionType.SingleValue);
 
             var verbose = cmd.Option(Constants.VerboseOption, Constants.VerboseDesc, CommandOptionType.NoValue);
+            verbose.ShowInHelpText = false;
+            var verbosity = cmd.Option(Constants.VerbosityOption, Constants.VerbosityDesc, CommandOptionType.SingleValue);
 
             cmd.HelpOption(Constants.HelpOption);
 
             cmd.OnExecuteAsync(async _ =>
             {
                 // Init logger
-                Util.SetVerbosity(log, verbose.HasValue());
+                Util.SetVerbosity(log, verbose.HasValue(), verbosity.Value());
 
                 var outputPath = output.HasValue() ? output.Value() : null;