+++ src/Sleet/InitAppCommand.cs
@@ -24,6 +24,8 @@ namespace Sleet
                 CommandOptionType.SingleValue);
 
             var verbose = cmd.Option(Constants.VerboseOption, Constants.VerboseDesc, CommandOptionType.NoValue);
+            verbose.ShowInHelpText = false;
+            var verbosity = cmd.Option(Constants.VerbosityOption, Constants.VerbosityDesc, CommandOptionType.SingleValue);
 
             cmd.HelpOption(Constants.HelpOption);
 
@@ -39,7 +41,7 @@ namespace Sleet
                 CmdUtils.VerifyRequiredOptions(required.ToArray());
 
                 // Init logger
-                Util.SetVerbosity(log, verbose.HasValue());
+                Util.SetVerbosity(log, verbose.HasValue(), verbosity.Value());
 
                 // Create a temporary folder for caching files during the operation.
                 using (var cache = new LocalCache())